Support location output parsed as CSV
Newer versions of Req will attempt to decode the body based on the
path name if the content-type is `application/octet-stream`. Since
this is being hardcoded to `*-manifest.csv`, that means it will
be decoded and fail the `String` function calls that were previously
used in `ReqAthena.S3.get_locations/2`.

This forces `decode_body: false` for this specific request since it
is expected to only be a newline delineated string