Use returned `DataManifestLocation` if available
Fixes #42

In some situations, the manifest file may be a slightly different path than
what is set in `output_locations` (reasons not fully known). This would
cause failures fetching the locations of results since the manifest would
not exist due to the hardcoded path.

This changes to use the `DataManifestLocation` key if it is returned
in the results rather than assuming the path